+++ CHANGELOG.md
@@ -9,6 +9,8 @@ levelAcceleration - defined by map car spawn delay (in ticks, on the start of th
 
 # BETA 0.2 SNAPSHOTS
 
+## v191101
+* Updated credits info - added ziputils library and CG license, updated sounds maker
 ## v191002
 + Added a short power description when hovering equip button or upgrade button.
 * New apperrance of Car Speed Increase anti-power.